refactor(algorithm): move scattered index implementations into dedicated subdirectories
Move brute_force, ivf, pyramid, sparse_index, and warp index
implementations from the top-level src/algorithm/ directory into their
own subdirectories, each with an independent CMakeLists.txt producing
a separate OBJECT library. This aligns them with the existing
hgraph/hnswlib/sindi pattern.

Additionally, merge the ivf_partition/ directory into ivf/, placing
all IVF-related code (core + partition strategies) under a single
directory to clarify IVF's code boundary.

Key changes:
- git mv index files into brute_force/, ivf/, pyramid/, sparse_index/,
  warp/ subdirectories
- Merge ivf_partition/ contents into ivf/; remove ivf_partition/
  directory and its OBJECT library target
- Add CMakeLists.txt for each new subdirectory
- Update top-level src/algorithm/CMakeLists.txt: add new
  add_subdirectory entries, remove ivf_partition, update ALGORITHM_LIBS
- Update all include paths across src/, tests/, and tools/ to reflect
  new file locations
- Run make fmt to ensure formatting compliance

No runtime behavior change; this is a pure structural refactor
affecting only the build system and include paths.
