+++ src/algorithm/CMakeLists.txt
@@ -1,3 +1,4 @@
+
 # Copyright 2024-present the vsag project
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,10 +14,14 @@
 # limitations under the License.
 
 
-add_subdirectory (ivf_partition)
 add_subdirectory (hnswlib)
 add_subdirectory (sindi)
 add_subdirectory (hgraph)
+add_subdirectory (bruteforce)
+add_subdirectory (ivf)
+add_subdirectory (pyramid)
+add_subdirectory (sparse_index)
+add_subdirectory (warp)
 
 file (GLOB ALGORITHM_SRCS "*.cpp")
 list (FILTER ALGORITHM_SRCS EXCLUDE REGEX "_test.cpp")
@@ -29,9 +34,13 @@ maybe_add_dependencies (algorithm antlr4)
 set (ALGORITHM_LIBS
         algorithm
         hnswlib
-        ivf_partition
         sindi
         hgraph
+        bruteforce
+        ivf
+        pyramid
+        sparse_index
+        warp
         PARENT_SCOPE)
 
 if (ENABLE_TESTS)
