+++ src/algorithm/inner_index_interface_test.cpp
@@ -18,8 +18,8 @@
 #include <memory>
 #include <sstream>
 
+#include "algorithm/bruteforce/bruteforce.h"
 #include "algorithm/hgraph/hgraph.h"
-#include "brute_force.h"
 #include "impl/allocator/safe_allocator.h"
 #include "index_common_param.h"
 #include "unittest.h"
