+++ src/algorithm/ivf/CMakeLists.txt
@@ -1,0 +1,28 @@
+
+# Copyright 2024-present the vsag project
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+set (IVF_SRCS
+        ivf.cpp
+        ivf_parameter.cpp
+        ivf_partition_strategy.cpp
+        ivf_partition_strategy_parameter.cpp
+        ivf_nearest_partition.cpp
+        gno_imi_partition.cpp
+        gno_imi_parameter.cpp
+)
+
+add_library (ivf OBJECT ${IVF_SRCS})
+target_link_libraries (ivf PRIVATE coverage_config vsag_src_common)
