+++ src/algorithm/ivf/ivf.cpp
@@ -23,6 +23,7 @@
 #include "attr/argparse.h"
 #include "attr/executor/executor.h"
 #include "datacell/flatten_interface.h"
+#include "gno_imi_partition.h"
 #include "impl/heap/standard_heap.h"
 #include "impl/inner_search_param.h"
 #include "impl/reorder/flatten_reorder.h"
@@ -30,8 +31,7 @@
 #include "index/index_impl.h"
 #include "index_feature_list.h"
 #include "inner_string_params.h"
-#include "ivf_partition/gno_imi_partition.h"
-#include "ivf_partition/ivf_nearest_partition.h"
+#include "ivf_nearest_partition.h"
 #include "query_context.h"
 #include "storage/serialization.h"
 #include "storage/stream_reader.h"
