+++ src/algorithm/pyramid/pyramid.cpp
