+++ src/algorithm/pyramid/pyramid_test.cpp
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "algorithm/pyramid.h"
+#include "pyramid.h"
 
 #include "unittest.h"
