+++ tests/test_warp.cpp
@@ -21,7 +21,7 @@
 #include <vector>
 
 #include "algorithm/inner_index_interface.h"
-#include "algorithm/warp.h"
+#include "algorithm/warp/warp.h"
 #include "data_type.h"
 #include "datacell/flatten_datacell_parameter.h"
 #include "framework/test_dataset.h"
