+++ tools/analyze_index/analyze_index.cpp
@@ -22,9 +22,9 @@
 #include <memory>
 
 #include "algorithm/hgraph/hgraph.h"
-#include "algorithm/ivf.h"
-#include "algorithm/pyramid.h"
-#include "algorithm/pyramid_zparameters.h"
+#include "algorithm/ivf/ivf.h"
+#include "algorithm/pyramid/pyramid.h"
+#include "algorithm/pyramid/pyramid_zparameters.h"
 #include "algorithm/sindi/sindi.h"
 #include "algorithm/sindi/sindi_parameter.h"
 #include "index/index_impl.h"
